Move library to Loki charm
