+++ tox.ini
@@ -8,8 +8,7 @@ envlist = lint, static, unit
 [vars]
 src_path = {toxinidir}/src/
 test_path = {toxinidir}/tests/
-lib_path = {toxinidir}/lib/charms/grafana_agent_k8s/v0
-all_path = {[vars]src_path} {[vars]test_path} {[vars]lib_path}
+all_path = {[vars]src_path} {[vars]test_path}
 
 [testenv]
 basepython = python3
